feat(MainRouter.tsx, App.tsx): transfer routing out of app into its separate place
