+++ frontend/package.json
@@ -11,6 +11,7 @@
   },
   "dependencies": {
     "@ag-grid-community/locale": "^33.3.2",
+    "@emotion/cache": "^11.14.0",
     "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.0",
     "@mui/icons-material": "^7.1.1",
@@ -18,6 +19,8 @@
     "ag-grid-react": "^33.3.2",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
+    "react-router-dom": "^7.6.2",
+    "stylis-plugin-rtl": "^2.1.1",
     "sweetalert2": "^11.22.0"
   },
   "devDependencies": {
